debugging

+++ src/testing/FileWriter.cu
@@ -125,10 +125,9 @@ std::optional<std::string> WriteResultsToFileDouble2(std::string file_name,
 }
 
 //Writes Benchmark results to file
-std::optional<std::string> WriteBenchResultsToFile(std::vector<double> run_avg,
-                                                   std::vector<double> run_sig,
-                                                   std::vector<int> fft_length,
-                                                   std::string sample_size){
+std::optional<std::string> WriteBenchResultsToFile(
+    std::vector<double> run_avg, std::vector<double> run_sig,
+    std::vector<long long> fft_length, std::string sample_size){
   std::ofstream myfile ("BenchResults_samples_" + sample_size + ".dat");
   if (myfile.is_open()) {
     for(int i=0; i<static_cast<int>(fft_length.size()); i++){
